+++ inc/Blocks/Image.php
@@ -34,6 +34,15 @@ class Image extends Block {
 		preg_match( '/src="([^"]+)"/', $block['originalContent'] ?? '', $matches );
 		$block['attributes']['url'] = esc_url( $matches[1] ?? '' );
 
+		// If it's not same site, get remote image.
+		if ( false === strpos( $block['attributes']['url'] ?? '', home_url() ) ) {
+			$remote_image = $this->get_remote_image( $block['attributes']['url'] ?? '' );
+
+			if ( ! is_wp_error( $remote_image ) ) {
+				$block['attributes']['url'] = $remote_image;
+			}
+		}
+
 		// Re-encode attributes correctly.
 		$block['attributes'] = wp_json_encode( $block['attributes'] );
 
@@ -55,5 +64,97 @@ class Image extends Block {
 		}
 
 		return $block;
+	}
+
+	/**
+	 * Import Image.
+	 *
+	 * This is specific to importing images from
+	 * an external or remote website.
+	 *
+	 * @since 1.3.0
+	 *
+	 * @param string $image_url Image URL.
+	 * @return string
+	 */
+	protected function get_remote_image( $image_url ): string {
+		if ( ! function_exists( 'download_url' ) ) {
+			require_once ABSPATH . 'wp-admin/includes/file.php';
+		}
+
+		if ( ! function_exists( 'wp_handle_sideload' ) ) {
+			require_once ABSPATH . 'wp-admin/includes/file.php';
+		}
+
+		if ( ! function_exists( 'wp_generate_attachment_metadata' ) ) {
+			require_once ABSPATH . 'wp-admin/includes/image.php';
+		}
+
+		// Download the file to a temporary location.
+		$tmp_file = download_url( sanitize_url( $image_url ) );
+
+		// Bail out, if wp_error.
+		if ( is_wp_error( $tmp_file ) ) {
+			error_log(
+				sprintf(
+					'Import error: %s, %s',
+					$tmp_file->get_error_message() ?? '',
+					$image_url
+				)
+			);
+
+			return $tmp_file;
+		}
+
+		// Get the filename + extension from the URL.
+		$url_filename = basename( parse_url( $image_url, PHP_URL_PATH ) );
+
+		// Build an array that resembles a PHP file upload.
+		$file = [
+			'name'     => $url_filename,
+			'type'     => mime_content_type( $tmp_file ),
+			'tmp_name' => $tmp_file,
+			'error'    => 0,
+			'size'     => filesize( $tmp_file ),
+		];
+
+		// Let WordPress handle the sideload.
+		$overrides = [
+			'test_form'   => false,
+			'test_size'   => true,
+			'test_upload' => true,
+		];
+
+		$results = wp_handle_sideload( $file, $overrides );
+
+		// Bail out, if upload error.
+		if ( isset( $results['error'] ) ) {
+			@unlink( $tmp_file );
+			$error_message = sprintf( 'Import error: %s', $results['error'] ?? '' );
+			error_log( $error_message );
+
+			return new WP_Error( 'cbtj-import-error', $error_message );
+		}
+
+		// Now create attachment post for the image.
+		$attachment = [
+			'post_mime_type' => $results['type'],
+			'post_title'     => sanitize_file_name( pathinfo( $url_filename, PATHINFO_FILENAME ) ),
+			'post_content'   => '',
+			'post_status'    => 'inherit',
+		];
+
+		$attach_id = wp_insert_attachment( $attachment, $results['file'] );
+
+		// Bail out, if wp_error.
+		if ( is_wp_error( $attach_id ) ) {
+			return $attach_id;
+		}
+
+		// Generate attachment metadata.
+		$metadata = wp_generate_attachment_metadata( $attach_id, $results['file'] );
+		wp_update_attachment_metadata( $attach_id, $metadata );
+
+		return wp_get_attachment_url( $attach_id );
 	}
 }