+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Feat: Added language translations for Japanese, Indonesia, Turkish, Polish, Dutch, Danish, Brazil, Portuguese.
 * Fix: Import issues with `core/list` and `core/list-item` blocks.
 * Feat: Add custom hook - `cbtj.innerBlocks`.
+* Feat: Implement Image import across websites.
 * Refactor: Replace `get_400_response` with `get_error_response`.
 * Chore: Add pull request template to repo.
 * Docs: Update README docs.